[BE] feat: 학습 시간 응답 값 실수형으로 변경

See merge request s12-final/S12P31C204!193

+++ spring/roll_model/src/main/java/com/ccc/roll_model/pipeline/domain/model/vo/ModelPerformanceSummary.java
@@ -2,7 +2,7 @@ package com.ccc.roll_model.pipeline.domain.model.vo;
 
 import lombok.Builder;
 
-public record ModelPerformanceSummary(String modelId, Integer learningDuration, Double result) {
+public record ModelPerformanceSummary(String modelId, Double learningDuration, Double result) {
 	@Builder
 	public ModelPerformanceSummary {
 	}
